finalize multiprocessing updates
